+++ src/Infrastructure/Dto/Commitment/CommitmentRequest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Domain\Commitment\Dto;
+namespace App\Infrastructure\Dto\Commitment;
 
-class CommitmentIn
+class CommitmentRequest
 {
     public string $orderId;
     public int $quantity;
